requirements.txt
astroid==2.5.2
certifi==2020.12.5
click==7.1.2
colorama==0.4.4
fastapi==0.63.0
h11==0.12.0
httpcore==0.12.3
httpx==0.17.1
idna==3.1
isort==5.8.0
lazy-object-proxy==1.6.0
loguru==0.5.3
mccabe==0.6.1
pydantic==1.8.1
pylint==2.7.4
PyYAML==5.4.1
rfc3986==1.4.0
sniffio==1.2.0
starlette==0.13.6
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
uvicorn==0.13.4
win32-setctime==1.0.3
wrapt==1.12.1
